requirements.txt
asgiref==3.8.1
Django==5.0.6
sqlparse==0.5.0
tzdata==2024.1
psycopg2-binary==2.9.9
pillow==10.3.0
djangorestframework==3.15.2
python-dotenv==1.0.1
black==24.4.2
click==8.1.7
colorama==0.4.6
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
django-filter==24.2
djangorestframework-simplejwt==5.3.1
pyjwt==2.8.0
coverage==7.5.4
drf-yasg==1.21.7
inflection==-0.5.1
pytz==2024.1
pyyaml==6.0.1
uritemplate==4.1.1
stripe==10.4.0







